+++ PuntoDeventa/PuntoDeventa/UI/CategoryProduct/Models/ProductSales.cs
@@ -33,7 +33,7 @@ namespace PuntoDeventa.UI.CategoryProduct.Models
         public string Name { get; set; }
         public string Description { get; set; }
 
-        public int BarCode { get; set; }
+        public long BarCode { get; set; }
 
         [Required(ErrorMessage = "El Sku es requerido.")]
         public int Sku { get; set; }
